WB_SmartBid/requirements.txt
pandas>=2.2.0
requests>=2.31.0
streamlit>=1.39.0
python-dotenv>=1.0.0
schedule>=1.2.0
plotly>=5.18.0
openpyxl>=3.1.0

